# api/swagger.yaml
#####################
# MAKE SURE YOU ARE MAKING CHANGES THROUGH https://app.swaggerhub.com/apis/guevarravirtual/cat-wrangler/1.0.0
#####################



swagger: '2.0'
info:
  version: 1.0.0
  title: cat wrangler
  description: For wrangling cats
schemes:
  - http
paths:
  /event:
    post:
      tags:
      - event
      summary: Creates a new event
      operationId: addEvent
      consumes:
      - application/json
      produces:
      - application/json
      parameters:
      - in: body
        name: body
        description: event object to be scheduled
        required: true
        schema:
          $ref: '#/definitions/Event'
      responses:
        405:
          description: Invalid input
    put:
      tags:
      - event
      summary: Update an existing event
      operationId: updateEvent
      consumes:
      - application/json
      produces:
      - application/json
      parameters:
      - in: body
        name: body
        description: event object that needs to be added to the store
        required: true
        schema:
          $ref: '#/definitions/Event'
      responses:
        400:
          description: Invalid ID supplied
        404:
          description: event not found
        405:
          description: Validation exception
  /event/{eventId}:
    get:
      tags:
      - event
      summary: Find event by ID
      description: Returns a single event
      operationId: getEventById
      produces:
      - application/json
      - application/xml
      parameters:
      - name: eventId
        in: path
        description: ID of event to return
        required: true
        type: integer
        format: int64
      responses:
        200:
          description: successful operation
          schema:
            $ref: '#/definitions/Event'
        400:
          description: Invalid ID supplied
        404:
          description: event not found
    post:
      tags:
      - event
      summary: Updates a event in the store with form data
      operationId: updateEventWithForm
      consumes:
      - application/x-www-form-urlencoded
      produces:
      - application/json
      - application/xml
      parameters:
      - name: eventId
        in: path
        description: ID of event that needs to be updated
        required: true
        type: integer
        format: int64
      - name: name
        in: formData
        description: Updated name of the event
        required: false
        type: string
      - name: status
        in: formData
        description: Updated status of the event
        required: false
        type: string
      responses:
        405:
          description: Invalid input
    delete:
      tags:
      - event
      summary: Deletes a event
      operationId: deleteEvent
      produces:
      - application/json
      - application/xml
      parameters:
      - name: api_key
        in: header
        required: false
        type: string
      - name: eventId
        in: path
        description: event id to delete
        required: true
        type: integer
        format: int64
      responses:
        400:
          description: Invalid ID supplied
        404:
          description: event not found
definitions:
  Event:
    type: object
    required:
    - name
    - startDateTime
    - creator
    properties:
      id:
        type: integer
        format: int64
      name:
        type: string
        example: board game christmas get together
      startDateTime:
        type: string
        format: date-time
        description: date time for when event starts
      creator:
        type: string
        description: User that created the event
        example: John Doe
  ApiResponse:
    type: object
    properties:
      code:
        type: integer
        format: int32
      type:
        type: string
      message:
        type: string